Add files via upload
This project describes about The Automatic Timetable Generator is a Python based software used to generate timetable automatically. It will help to manage all the periods automatically and will be helpful for faculty to get timetable in their phone by using application. It mainly addresses with Genetic algorithm that follows the idea of SURVIVAL OF THE FITTEST- Better and better solutions evolve from previous generations until a near optimal solution is obtained. It focusses on optimization of resources i.e. teachers, labs and rooms etc. Provide a facility for everyone to view timetable.